+++ changelog.config.js
@@ -1,28 +1,10 @@
 module.exports = {
     disableEmoji: false,
     format: '{type}{scope}: {emoji}{subject}',
-    list: [
-        'test',
-        'feat',
-        'fix',
-        'chore',
-        'docs',
-        'refactor',
-        'style',
-        'ci',
-        'perf',
-    ],
+    list: ['test', 'feat', 'fix', 'chore', 'docs', 'refactor', 'style', 'ci', 'perf'],
     maxMessageLength: 64,
     minMessageLength: 3,
-    questions: [
-        'type',
-        'scope',
-        'subject',
-        'body',
-        'breaking',
-        'issues',
-        'lerna',
-    ],
+    questions: ['type', 'scope', 'subject', 'body', 'breaking', 'issues', 'lerna'],
     scopes: [],
     types: {
         chore: {
@@ -56,8 +38,7 @@ module.exports = {
             value: 'perf',
         },
         refactor: {
-            description:
-                'A code change that neither fixes a bug or adds a feature',
+            description: 'A code change that neither fixes a bug or adds a feature',
             emoji: '(💡)',
             value: 'refactor',
         },
@@ -67,8 +48,7 @@ module.exports = {
             value: 'release',
         },
         style: {
-            description:
-                'Markup, white-space, formatting, missing semi-colons...',
+            description: 'Markup, white-space, formatting, missing semi-colons...',
             emoji: '(💄)',
             value: 'style',
         },
@@ -80,8 +60,7 @@ module.exports = {
         messages: {
             type: "Select the type of change that you're committing:",
             customScope: 'Select the scope this component affects:',
-            subject:
-                'Write a short, imperative mood description of the change:\n',
+            subject: 'Write a short, imperative mood description of the change:\n',
             body: 'Provide a longer description of the change:\n ',
             breaking: 'List any breaking changes:\n',
             footer: 'Issues this commit closes, e.g #123:',