+++ packages/contexts/src/dialogflow/index.js
@@ -5,8 +5,7 @@ const DialogFlowClass = require('./dialogflow.class')
  * @param {*} args
  * @returns
  */
-const createBotDialog = async ({ database, provider }) =>
-    new DialogFlowClass(database, provider)
+const createBotDialog = async ({ database, provider }) => new DialogFlowClass(database, provider)
 
 module.exports = {
     createBotDialog,